NFC: move MIDI devices to separate files
