+++ src/deluge/io/midi/midi_device.cpp
@@ -352,225 +352,3 @@ bool MIDIPort::worthWritingToFile() {
 	return (mpeLowerZoneLastMemberChannel || mpeUpperZoneLastMemberChannel != 15);
 }
 
-void MIDIDeviceUSB::connectedNow(int32_t midiDeviceNum) {
-	connectionFlags |= (1 << midiDeviceNum);
-	needsToSendMCMs = 2;
-}
-
-void MIDIDeviceUSB::sendMCMsNowIfNeeded() {
-	if (needsToSendMCMs) {
-		needsToSendMCMs--;
-		if (!needsToSendMCMs) {
-			sendAllMCMs();
-		}
-	}
-}
-
-void MIDIDeviceUSB::sendMessage(uint8_t statusType, uint8_t channel, uint8_t data1, uint8_t data2) {
-	if (!connectionFlags) {
-		return;
-	}
-
-	int32_t ip = 0;
-
-	uint32_t fullMessage = setupUSBMessage(statusType, channel, data1, data2);
-
-	for (int32_t d = 0; d < MAX_NUM_USB_MIDI_DEVICES; d++) {
-		if (connectionFlags & (1 << d)) {
-			ConnectedUSBMIDIDevice* connectedDevice = &connectedUSBMIDIDevices[ip][d];
-			connectedDevice->bufferMessage(fullMessage);
-		}
-	}
-}
-
-int32_t MIDIDeviceUSB::sendBufferSpace() {
-	int32_t ip = 0;
-	ConnectedUSBMIDIDevice* connectedDevice = NULL;
-
-	// find the connected device for this specific device. Note that virtual
-	// port number is specified as part of the message, implemented below.
-	for (int32_t d = 0; d < MAX_NUM_USB_MIDI_DEVICES; d++) {
-		if (connectionFlags & (1 << d)) {
-			connectedDevice = &connectedUSBMIDIDevices[ip][d];
-			break;
-		}
-	}
-
-	if (!connectedDevice) {
-		return 0;
-	}
-
-	return connectedDevice->sendBufferSpace();
-}
-
-extern bool developerSysexCodeReceived;
-
-void MIDIDeviceUSB::sendSysex(const uint8_t* data, int32_t len) {
-	if (len < 6 || data[0] != 0xf0 || data[len - 1] != 0xf7) {
-		return;
-	}
-
-	int32_t ip = 0;
-	ConnectedUSBMIDIDevice* connectedDevice = NULL;
-
-	// find the connected device for this specific device. Note that virtual
-	// port number is specified as part of the message, implemented below.
-	for (int32_t d = 0; d < MAX_NUM_USB_MIDI_DEVICES; d++) {
-		if (connectionFlags & (1 << d)) {
-			connectedDevice = &connectedUSBMIDIDevices[ip][d];
-			break;
-		}
-	}
-
-	if (!connectedDevice) {
-		return;
-	}
-
-	int32_t pos = 0;
-
-	// While we are standardizing on using the 4 byte Synthstrom Deluge in our messages, it is possible that client
-	// programs have not been updated yet, so if we get a SysEx request with 0x7D as the ID, we respond in-kind. This
-	// involves skipping the the first 5 bytes and sending 0xF0, 0x7D instead. Since the USB driver batches things in 3
-	// byte groups, we must include the first byte after the address in that first group.
-	if (developerSysexCodeReceived && (data[1] != 0x7D)) {
-		// Since the message ends with 0xF7, we can assume that data[5] does exist.
-		// fake 0xF0, 0x7D, data[5] for first send
-		uint32_t packed = ((uint32_t)data[5] << 24) | 0x007DF004 | (portNumber << 4);
-		connectedDevice->bufferMessage(packed);
-		pos = 6;
-	}
-
-	while (pos < len) {
-		int32_t status, byte0 = 0, byte1 = 0, byte2 = 0;
-		byte0 = data[pos];
-		if (len - pos > 3) {
-			status = 0x4; // sysex start or continue
-			byte1 = data[pos + 1];
-			byte2 = data[pos + 2];
-			pos += 3;
-		}
-		else {
-			status = 0x4 + (len - pos); // sysex end with N bytes
-			if ((len - pos) > 1) {
-				byte1 = data[pos + 1];
-			}
-			if ((len - pos) > 2) {
-				byte2 = data[pos + 2];
-			}
-			pos = len;
-		}
-		status |= (portNumber << 4);
-		uint32_t packed = ((uint32_t)byte2 << 24) | ((uint32_t)byte1 << 16) | ((uint32_t)byte0 << 8) | status;
-		connectedDevice->bufferMessage(packed);
-	}
-}
-
-void MIDIDeviceUSBHosted::writeReferenceAttributesToFile(Serializer& writer) {
-	writer.writeAttribute("name", name.get());
-	writer.writeAttributeHex("vendorId", vendorId, 4);
-	writer.writeAttributeHex("productId", productId, 4);
-}
-
-void MIDIDeviceUSBHosted::writeToFlash(uint8_t* memory) {
-	*(uint16_t*)memory = vendorId;
-	*(uint16_t*)(memory + 2) = productId;
-}
-
-char const* MIDIDeviceUSBHosted::getDisplayName() {
-	return name.get();
-}
-
-void MIDIDeviceUSBHosted::callHook(Hook hook) {
-	switch (hook) {
-	case Hook::HOOK_ON_CONNECTED:
-		hookOnConnected();
-		break;
-	case Hook::HOOK_ON_CHANGE_ROOT_NOTE:
-		hookOnChangeRootNote();
-		break;
-	case Hook::HOOK_ON_CHANGE_SCALE:
-		hookOnChangeScale();
-		break;
-	case Hook::HOOK_ON_ENTER_SCALE_MODE:
-		hookOnEnterScaleMode();
-		break;
-	case Hook::HOOK_ON_EXIT_SCALE_MODE:
-		hookOnExitScaleMode();
-		break;
-	case Hook::HOOK_ON_MIDI_LEARN:
-		hookOnMIDILearn();
-		break;
-	case Hook::HOOK_ON_RECALCULATE_COLOUR:
-		hookOnRecalculateColour();
-		break;
-	case Hook::HOOK_ON_TRANSITION_TO_ARRANGER_VIEW:
-		hookOnTransitionToArrangerView();
-		break;
-	case Hook::HOOK_ON_TRANSITION_TO_CLIP_VIEW:
-		hookOnTransitionToClipView();
-		break;
-	case Hook::HOOK_ON_TRANSITION_TO_SESSION_VIEW:
-		hookOnTransitionToSessionView();
-		break;
-	case Hook::HOOK_ON_WRITE_HOSTED_DEVICE_TO_FILE:
-		hookOnWriteHostedDeviceToFile();
-		break;
-	default:
-		break;
-	}
-}
-
-void MIDIDeviceUSBUpstream::writeReferenceAttributesToFile(Serializer& writer) {
-	// Same line. Usually the user wouldn't have default velocity sensitivity set for their computer.
-	writer.writeAttribute("port", portNumber ? "upstreamUSB2" : "upstreamUSB", false);
-}
-
-void MIDIDeviceUSBUpstream::writeToFlash(uint8_t* memory) {
-	D_PRINTLN("writing to flash port  %d  into ", portNumber);
-	*(uint16_t*)memory = portNumber ? VENDOR_ID_UPSTREAM_USB2 : VENDOR_ID_UPSTREAM_USB;
-}
-
-char const* MIDIDeviceUSBUpstream::getDisplayName() {
-	switch (portNumber) {
-	case 0:
-		return deluge::l10n::get(deluge::l10n::String::STRING_FOR_UPSTREAM_USB_PORT_1);
-	case 1:
-		return deluge::l10n::get(deluge::l10n::String::STRING_FOR_UPSTREAM_USB_PORT_2);
-	case 2:
-		return deluge::l10n::get(deluge::l10n::String::STRING_FOR_UPSTREAM_USB_PORT_3_SYSEX);
-	default:
-		return "";
-	}
-}
-
-void MIDIDeviceDINPorts::writeReferenceAttributesToFile(Serializer& writer) {
-	// Same line. Usually the user wouldn't have default velocity sensitivity set
-	writer.writeAttribute("port", "din", false);
-}
-
-void MIDIDeviceDINPorts::writeToFlash(uint8_t* memory) {
-	*(uint16_t*)memory = VENDOR_ID_DIN;
-}
-
-char const* MIDIDeviceDINPorts::getDisplayName() {
-	return deluge::l10n::get(deluge::l10n::String::STRING_FOR_DIN_PORTS);
-}
-
-void MIDIDeviceDINPorts::sendMessage(uint8_t statusType, uint8_t channel, uint8_t data1, uint8_t data2) {
-	midiEngine.sendSerialMidi(statusType, channel, data1, data2);
-}
-
-int32_t MIDIDeviceDINPorts::sendBufferSpace() {
-	return uartGetTxBufferSpace(UART_ITEM_MIDI);
-}
-
-void MIDIDeviceDINPorts::sendSysex(const uint8_t* data, int32_t len) {
-	if (len < 3 || data[0] != 0xf0 || data[len - 1] != 0xf7) {
-		return;
-	}
-
-	// NB: beware of MIDI_TX_BUFFER_SIZE
-	for (int32_t i = 0; i < len; i++) {
-		bufferMIDIUart(data[i]);
-	}
-}
