+++ src/deluge/io/midi/midi_device.h
@@ -141,107 +141,3 @@ protected:
 	// These only go into SETTINGS/MIDIDevices.XML
 	void writeDefinitionAttributesToFile(Serializer& writer);
 };
-
-class MIDIDeviceUSB : public MIDIDevice {
-public:
-	MIDIDeviceUSB(uint8_t portNum = 0) {
-		portNumber = portNum;
-		needsToSendMCMs = 0;
-	}
-	void sendMessage(uint8_t statusType, uint8_t channel, uint8_t data1, uint8_t data2);
-	void sendSysex(const uint8_t* data, int32_t len) override;
-	int32_t sendBufferSpace() override;
-	void connectedNow(int32_t midiDeviceNum);
-	void sendMCMsNowIfNeeded();
-	uint8_t needsToSendMCMs;
-	uint8_t portNumber;
-};
-
-class MIDIDeviceUSBHosted : public MIDIDeviceUSB {
-public:
-	MIDIDeviceUSBHosted() {}
-	void writeReferenceAttributesToFile(Serializer& writer);
-	void writeToFlash(uint8_t* memory);
-	char const* getDisplayName();
-
-	// Add new hooks to the below list.
-
-	// Gets called once for each freshly connected hosted device.
-	virtual void hookOnConnected(){};
-
-	// Gets called when something happens that changes the root note
-	virtual void hookOnChangeRootNote(){};
-
-	// Gets called when something happens that changes the scale/mode
-	virtual void hookOnChangeScale(){};
-
-	// Gets called when entering Scale Mode in a clip
-	virtual void hookOnEnterScaleMode(){};
-
-	// Gets called when exiting Scale Mode in a clip
-	virtual void hookOnExitScaleMode(){};
-
-	// Gets called when learning/unlearning a midi device to a clip
-	virtual void hookOnMIDILearn(){};
-
-	// Gets called when recalculating colour in clip mode
-	virtual void hookOnRecalculateColour(){};
-
-	// Gets called when transitioning to ArrangerView
-	virtual void hookOnTransitionToArrangerView(){};
-
-	// Gets called when transitioning to ClipView
-	virtual void hookOnTransitionToClipView(){};
-
-	// Gets called when transitioning to SessionView
-	virtual void hookOnTransitionToSessionView(){};
-
-	// Gets called when hosted device info is saved to XML (usually after changing settings)
-	virtual void hookOnWriteHostedDeviceToFile(){};
-
-	// Add an entry to this enum if adding new virtual hook functions above
-	enum class Hook {
-		HOOK_ON_CONNECTED = 0,
-		HOOK_ON_CHANGE_ROOT_NOTE,
-		HOOK_ON_CHANGE_SCALE,
-		HOOK_ON_ENTER_SCALE_MODE,
-		HOOK_ON_EXIT_SCALE_MODE,
-		HOOK_ON_MIDI_LEARN,
-		HOOK_ON_RECALCULATE_COLOUR,
-		HOOK_ON_TRANSITION_TO_ARRANGER_VIEW,
-		HOOK_ON_TRANSITION_TO_CLIP_VIEW,
-		HOOK_ON_TRANSITION_TO_SESSION_VIEW,
-		HOOK_ON_WRITE_HOSTED_DEVICE_TO_FILE
-	};
-
-	// Ensure to add a case to this function in midi_device.cpp when adding new hooks
-	void callHook(Hook hook);
-
-	uint16_t vendorId;
-	uint16_t productId;
-
-	bool freshly_connected = true; // Used to trigger hookOnConnected from the input loop
-
-	String name;
-};
-
-class MIDIDeviceUSBUpstream final : public MIDIDeviceUSB {
-public:
-	MIDIDeviceUSBUpstream(uint8_t portNum = 0) : MIDIDeviceUSB(portNum) {}
-	void writeReferenceAttributesToFile(Serializer& writer);
-	void writeToFlash(uint8_t* memory);
-	char const* getDisplayName();
-};
-
-class MIDIDeviceDINPorts final : public MIDIDevice {
-public:
-	MIDIDeviceDINPorts() {
-		connectionFlags = 1; // DIN ports are always connected
-	}
-	void writeReferenceAttributesToFile(Serializer& writer);
-	void writeToFlash(uint8_t* memory);
-	char const* getDisplayName();
-	void sendMessage(uint8_t statusType, uint8_t channel, uint8_t data1, uint8_t data2);
-	void sendSysex(const uint8_t* data, int32_t len) override;
-	int32_t sendBufferSpace() override;
-};
